pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "probraw"
dynamic = ["version"]
description = "Pipeline reproducible para RAW, deteccion de carta y perfilado ICC de camara"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
authors = [
  {name = "Probatia Forensics SL y Asociacion Espanola de Imagen Cientifica y Forense"}
]
classifiers = [
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
  "numpy>=2.0",
  "PyYAML>=6.0",
  "opencv-python-headless>=4.10",
  "tifffile>=2025.1",
  "Pillow>=10.0",
  "colour-science>=0.4.7",
  "scipy>=1.13",
  "rawpy>=0.26",
  "cryptography>=41.0",
]

[project.optional-dependencies]
gui = [
  "PySide6>=6.8",
]
dev = [
  "pytest>=8.0",
  "py-spy>=0.3.14",
]
installer = [
  "pyinstaller>=6.10",
]
c2pa = [
  "c2pa-python>=0.32",
]

[project.scripts]
probraw = "probraw.cli:main"
probraw-ui = "probraw.gui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "probraw.version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"probraw.resources" = ["references/*.json", "icons/*.svg", "icons/*.png", "icons/*.ico", "locales/*.qm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
